@@ -8,10 +8,9 @@
 		<plugins>
 			<plugin>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>2.4</version>
+				<version>3.13.0</version>
 				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
+					<release>11</release>
 					<compilerArgument>-AaddGeneratedAnnotation=true</compilerArgument>
 					<compilerArguments>
 						<AaddGeneratedAnnotation>false</AaddGeneratedAnnotation>
