@@ -74,13 +74,13 @@ public class JavaProjectConversionParticipant extends AbstractProjectConversionP
 
   private static final String DEFAULT_TEST_RESOURCES = "src/test/resources"; //$NON-NLS-1$
 
-  private static final String DEFAULT_JAVA_VERSION = "1.5"; //$NON-NLS-1$
+  private static final String DEFAULT_JAVA_VERSION = "1.8"; //$NON-NLS-1$
 
   private static final String COMPILER_GROUP_ID = "org.apache.maven.plugins"; //$NON-NLS-1$
 
   private static final String COMPILER_ARTIFACT_ID = "maven-compiler-plugin"; //$NON-NLS-1$
 
-  private static final String DEFAULT_COMPILER_VERSION = "3.8.1"; //$NON-NLS-1$
+  private static final String DEFAULT_COMPILER_VERSION = "3.13.0"; //$NON-NLS-1$
 
   private static final String TARGET_KEY = "target"; //$NON-NLS-1$
 
